Testspassed

+++ src/units/Unit.java
@@ -36,7 +36,7 @@ public abstract class Unit {
     this.currentSoldierCount = currentSoldierCount;
   }
   public void decCurrentSoldierCount(){
-    currentSoldierCount -= currentSoldierCount * 0.1;
+    currentSoldierCount -= (int)(currentSoldierCount * 0.1);
   }
 
   public double getIdleUpkeep() {
@@ -78,7 +78,7 @@ public abstract class Unit {
     if (this.getParentArmy().equals(target.getParentArmy())) {
       throw new FriendlyFireException();
     } else {
-      target.currentSoldierCount -= this.currentSoldierCount * this.unitFactor(target, level);
+      target.currentSoldierCount -= (int)(this.currentSoldierCount * this.unitFactor(target, level));
       if (target.currentSoldierCount<0){
         target.currentSoldierCount = 0;
       }
